Break apart nagent.h/.cpp
