+++ salalib/agents/agents.pri
@@ -1,0 +1,13 @@
+HEADERS += \
+    $$PWD/agent.h \
+    $$PWD/agentprogram.h \
+    $$PWD/agentset.h \
+    $$PWD/agentengine.h \
+    $$PWD/agentga.h \
+    $$PWD/agenthelpers.h
+SOURCES += \
+    $$PWD/agent.cpp \
+    $$PWD/agentprogram.cpp \
+    $$PWD/agentset.cpp \
+    $$PWD/agentengine.cpp \
+    $$PWD/agentga.cpp
